fix(auth): fix default category creation for new users on PostgreSQL
- Add UserService and CategoryService to auth server for category creation
- Share auth server instance between gRPC and REST handlers
- Fix PostgreSQL enum type mismatch by storing enums as int32 in models
- Convert Category.Type, Wallet.Type, Wallet.Status from enum to int32
- Add explicit type casting for all enum comparisons and assignments
- Add helper methods for enum conversion in models

Fixes issue where default categories weren't created during user
registration due to auth handlers creating new server instances and
GORM sending enum names instead of integer values to PostgreSQL.

Changes are backward compatible with proper enum-to-int32 conversion.

+++ src/go-backend/domain/repository/category_repository.go
@@ -193,7 +193,7 @@ func (r *categoryRepository) CreateDefaultCategories(ctx context.Context, userID
 		category := &models.Category{
 			UserID: userID,
 			Name:   name,
-			Type:   v1.CategoryType_CATEGORY_TYPE_EXPENSE,
+			Type:   int32(v1.CategoryType_CATEGORY_TYPE_EXPENSE),
 		}
 		if err := r.Create(ctx, category); err != nil {
 			return apperrors.NewInternalErrorWithCause("failed to create default expense category", err)
@@ -205,7 +205,7 @@ func (r *categoryRepository) CreateDefaultCategories(ctx context.Context, userID
 		category := &models.Category{
 			UserID: userID,
 			Name:   name,
-			Type:   v1.CategoryType_CATEGORY_TYPE_INCOME,
+			Type:   int32(v1.CategoryType_CATEGORY_TYPE_INCOME),
 		}
 		if err := r.Create(ctx, category); err != nil {
 			return apperrors.NewInternalErrorWithCause("failed to create default income category", err)
@@ -232,7 +232,7 @@ func (r *categoryRepository) GetByNameAndType(ctx context.Context, userID int32,
 	newCategory := &models.Category{
 		UserID: userID,
 		Name:   name,
-		Type:   categoryType,
+		Type:   int32(categoryType),
 	}
 
 	if err := r.Create(ctx, newCategory); err != nil {
